Put ArenaObject class into own header -- it's useful
